Adds a replacement for RelativeRotationDirection connectivity and tests style in ResistorBundle
