+++ src/main/java/org/eln2/mc/common/cells/foundation/CellPos.kt
@@ -1,3 +1,5 @@
 package org.eln2.mc.common.cells.foundation
 
+import org.eln2.mc.common.space.LocationDescriptor
+
 data class CellPos(val descriptor: LocationDescriptor)
